fix: Remove web page script from Website Theme
